+++ ambra_sdk/__init__.py
@@ -1,8 +1,8 @@
-__version__ = '3.20.8.0-2'
+__version__ = '3.21.1.0-1'
 
 # from footer in api.html
-API_VERSION = 'LBL0022 v42.0 2020-12-02'
+API_VERSION = 'LBL0022 v44.0 2021-02-24'
 # equal for API_VERSION based on dictionary.html
-MODELS_VERSION = 'LBL0022 v42.0 2020-12-02'
+MODELS_VERSION = 'LBL0022 v44.0 2021-02-24'
 # from storage_api.html
-STORAGE_VERSION = 'LBL0038 v11.0 2020-10-07'
+STORAGE_VERSION = 'LBL0038 v13.0 2021-01-13'