+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [3.21.1.0-1] - 2021-01-13
+### Added
+- Service public api support: LBL0022 v44.0 2021-02-24
+- Service models: LBL0022 v44.0 2021-02-24
+- Storage public api support: LBL0038 v13.0 2021-01-13
+- api.Addon.Study.wait_job method
+- api.Addon.Study.wait_anonymize_and_wait method
+- api.Addon.Study.wait_anonymize_and_get method
+- Support filtering using datetime with timezones
+
 ## [3.20.8.0-2] - 2020-11-18
 ### Added
 - Service public api support: LBL0022 v42.0 2020-12-02